pyproject.toml
[project]
name = "ecoevojobs"
version = "0.1.0"
description = "Quick analysis of the ecoevojobs data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cartopy>=0.24.1",
    "fastexcel>=0.12.0",
    "gdown>=5.2.0",
    "jupyterlab>=4.3.3",
    "lxml>=5.3.0",
    "matplotlib>=3.9.3",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "polars>=1.17.1",
    "pyarrow>=18.1.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.0",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "snakemake>=8.25.5",
    "statsmodels>=0.14.4",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.1",
    "us>=3.2.0",
    "xlsx2csv>=0.8.4",
]
